dd Gaudd2DInt to sherpa_ class directly and not in the test
